+++ src/main/java/com/qin/catcat/unite/param/PassApproveParam.java
@@ -12,5 +12,6 @@ import lombok.Data;
 @Data
 public class PassApproveParam {
     // 帖子ID
-    private Long postId;
+    private Integer postId;
 }
+
