+++ src/main/java/com/qin/catcat/unite/param/RefuseApproveParam.java
@@ -12,5 +12,6 @@ import lombok.Data;
 @Data
 public class RefuseApproveParam {
     // 帖子ID
-    private Long postId;
+    private Integer postId;
 }
+
